restructuring where blockbinary lives
