+++ tests/native/block_sub.cpp
@@ -8,7 +8,7 @@
 
 // minimum set of include files to reflect source code dependencies
 #include "universal/native/integers.hpp" // for to_binary(int)
-#include "universal/native/blockbinary.hpp"
+#include "universal/blockbin/blockbinary.hpp"
 // test helpers, such as, ReportTestResults
 #include "../utils/test_helpers.hpp"
 #include "../utils/blockbinary_helpers.hpp"
